Create Q7 TimeConverter
Write a program that converts from 24-hour notation to 12-hour notation.

For example, it should convert 14:25 to 2:25 PM. The input is given as two

integers. There should be at least three functions, one for input, one to do the

conversion, and one for output. Record the AM/PM information as a value of

type char, ‘A’ for AM and ‘P’ for PM. Thus, the function for doing the conversions

will have a call-by-reference formal parameter of type char to record

whether it is AM or PM. (The function will have other parameters as well.)

Include a loop that lets the user repeat this computation for new input values

again and again until the user says he or she wants to end the program.