Create Q9 Average
Write a program that calculates the average of a group of test scores, where the lowest

score in the group is dropped. It should use the following functions:

• void getScre(int &,int &,int &,int &,int &) should ask the user for a test scores, store it in a reference

parameter variable. This function should be called by main once.  Check for valid values with range

of 1 to 100.

• float calcAvg(int,int,int,int,int) should calculate and display the average of the four highest

scores. This function should be called just once by main, and should be passed

the five scores.

• int fndLwst(int,int,int,int,int) should find and return the lowest of the five scores passed to

it. It should be called by calcAvg, which uses the function to determine which

of the five scores to drop.